backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
langchain-groq
